+++ NyaaTools/operators/NyaaToolsAvatarMergeExport.py
@@ -4,6 +4,7 @@ from bpy.props import StringProperty
 
 from ..mesh.cleanup_mesh import cleanup_mesh
 from ..consts import EXPORT_COLLECTION, PROP_AVATAR_EXPORT_PATH
+from ..common.deselect_all import deselect_all
 from ..common.selection_add import selection_add
 from ..common.select_collection import select_collection
 from ..common.get_prop import get_prop
@@ -15,34 +16,48 @@ from ..avatar.get_avatar_armature import get_avatar_armature
 
 class NyaaToolsAvatarMergeExport(bpy.types.Operator):
     """Merge and export the avatar. If you have an export path defined, it will export there"""
+
     bl_idname = "nyaa.avatar_merge_export"
     bl_label = "Merge & Export"
     bl_options = {"REGISTER", "UNDO"}
 
-    avatar_name: StringProperty(
-        name="Avatar Name",
-        default=""
-    )
+    avatar_name: StringProperty(name="Avatar Name", default="")
+    # TODO: additional options
 
     def execute(self, context):
         try:
-            perform_merge_export(self.avatar_name)
+            if not self.avatar_name:
+                self.report({"ERROR"}, "Specify an avatar name.")
+                return {"CANCELLED"}
+
+            perform_merge_export(
+                self.avatar_name,
+                # TODO: additional options
+            )
+            return {"FINISHED"}
         except Exception as error:
             print(traceback.format_exc())
             self.report({"ERROR"}, str(error))
             return {"CANCELLED"}
 
-        return {"FINISHED"}
-
 
 def perform_merge_export(avatar_name):
+    def debug_print(*msgs):
+        print("   ", *msgs)
+        return
+
+    bpy.ops.object.mode_set(mode="OBJECT")
+
     armature = get_avatar_armature(avatar_name)
 
-    # Create "Export" collection
-    export_collection = bpy.data.collections.new("Export")
+    # Show all collections
+    for col in bpy.context.scene.collection.children:
+        col.hide_viewport = False
 
-    bpy.context.scene.collection.children.link(export_collection)
-    export_collection.color_tag = "COLOR_01"
+    # Create export collection
+    export_col = bpy.data.collections.new("__Export Temp__")
+    bpy.context.scene.collection.children.link(export_col)
+    export_col.color_tag = "COLOR_01"
 
     # Rename all objects to avoid collisions
     for obj in bpy.data.objects:
@@ -53,11 +68,17 @@ def perform_merge_export(avatar_name):
     armature.data.name = "Armature"
     armature.parent = None
 
+    # Link to soon-to-be Export collection
     try:
         selection_add(armature)
-        bpy.ops.object.transform_apply(
-            location=True, rotation=True, scale=True)
-        bpy.data.collections[EXPORT_COLLECTION].objects.link(armature)
+        bpy.ops.object.transform_apply(location=True, rotation=True, scale=True)
+        bpy.data.collections["__Export Temp__"].objects.link(armature)
+    except:
+        None
+
+    # Unlink from Scene Collection
+    try:
+        bpy.context.scene.collection.objects.unlink(armature)
     except:
         None
 
@@ -70,35 +91,37 @@ def perform_merge_export(avatar_name):
             layers = get_avatar_layers(mesh)
             for path_avatar_name, path_layer_name in layers:
                 if avatar_name == path_avatar_name:
-                    merge_onto_avatar_layer(
-                        path_layer_name, meshName, armature)
+                    merge_onto_avatar_layer(path_layer_name, meshName, armature)
 
         else:
             print("    BUG: Mesh doesn't exist, skipping for now:  " + meshName)
 
     # Cleanup temp objects
     for obj in bpy.data.objects:
-        if (obj.name.startswith("____")):
+        if obj.name.startswith("____"):
             bpy.data.objects.remove(obj)
 
     # Optimize meshes
     for obj in bpy.data.objects:
-        if (obj.type == "MESH"):
+        if obj.type == "MESH":
             cleanup_mesh(obj)
 
     for col in bpy.context.scene.collection.children:
-        if (col.name != "Export"):
+        if col.name != "__Export Temp__":
             bpy.data.collections.remove(col)
 
     # Export
+    export_col.name = EXPORT_COLLECTION
     select_collection(EXPORT_COLLECTION)
+
+    deselect_all()
+    selection_add(armature)
 
     if get_prop(armature, PROP_AVATAR_EXPORT_PATH):
         path = get_prop(armature, PROP_AVATAR_EXPORT_PATH)
-        if (path):
-            if (path[-1] == "/" or path[-1] == "\\"):
-                path = bpy.path.abspath(
-                    "//" + path + path_avatar_name + ".fbx")
+        if path:
+            if path[-1] == "/" or path[-1] == "\\":
+                path = bpy.path.abspath("//" + path + path_avatar_name + ".fbx")
             else:
                 path = bpy.path.abspath("//" + path)
 
@@ -107,7 +130,6 @@ def perform_merge_export(avatar_name):
             check_existing=False,
             filter_glob="*.fbx",
             use_active_collection=True,
-
             apply_scale_options="FBX_SCALE_UNITS",
             axis_forward="-Y",
             axis_up="Z",
@@ -116,7 +138,6 @@ def perform_merge_export(avatar_name):
             # primary_bone_axis="X",
             # secondary_bone_axis="-Y",
             add_leaf_bones=False,
-
             bake_anim=False
             # bake_anim_use_all_actions=False,
         )